feat: Updated content and added new posts
- Removed the title and description from the main index page.
- Created a new post titled "日記をはじめた" in the "diary" category.
- Added a new index page for the year 2023 with a brief description.

These changes were made to make the diary site more organized and
user-friendly. The new post shares about starting a diary website and
the choice of using a blog format instead of Twitter. It also mentions
the preference for having full control over the content and the decision
to use GitHub Pages for hosting. The deleted post "やあ" was removed as it
was no longer needed.